+++ src/app/cli.ts
@@ -2,8 +2,8 @@
 import Pastel from 'pastel';
 import { enableMapSet } from 'immer';
 
-
 enableMapSet()
+
 const app = new Pastel({
     importMeta: import.meta,
     name: '🕸️☁️ Uix CLI',